Updated common.rs to use ast_serde
